requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pillow==10.4.0
requests==2.32.3
numpy==1.26.4
tensorflow-cpu==2.16.1
